+++ layers/GatoGraphQLForWP/plugins/gatographql/src/PluginSkeleton/AbstractPlugin.php
@@ -18,6 +18,7 @@ use PoP\Root\Helpers\ClassHelpers;
 use PoP\Root\Helpers\ScopingHelpers;
 use PoP\Root\Module\ModuleInterface;
 
+use function add_action;
 use function get_option;
 use function flush_rewrite_rules;
 
@@ -449,7 +450,7 @@ abstract class AbstractPlugin implements PluginInterface
          *
          * @see layers/GatoGraphQLForWP/plugins/gatographql/src/Services/Taxonomies/AbstractTaxonomy.php
          */
-        \add_action(
+        add_action(
             'init',
             function (): void {
                 $this->maybeInstallPluginSetupData();
@@ -465,6 +466,19 @@ abstract class AbstractPlugin implements PluginInterface
     public function isLicenseJustActivated(): void
     {
         // Override
+    }
+
+    /**
+     * Method to override.
+     *
+     * Execute actions when the depended-upon plugin or theme's
+     * status has changed (activated/deactivated).
+     *
+     * @param string[] $pluginFilesOrThemeSlugsWithStatusChange The plugin files or theme slugs with status change
+     */
+    public function dependedUponPluginOrThemeStatusJustChanged(array $pluginFilesOrThemeSlugsWithStatusChange): void
+    {
+        // Override this method
     }
 
     /**
@@ -486,7 +500,7 @@ abstract class AbstractPlugin implements PluginInterface
          *
          * @see layers/GatoGraphQLForWP/plugins/gatographql/src/Services/Taxonomies/AbstractTaxonomy.php
          */
-        \add_action(
+        add_action(
             'init',
             function () use ($previousVersion): void {
                 // The version could contain the hash commit. Remove it!