+++ layers/GatoGraphQLForWP/plugins/gatographql/src/PluginSkeleton/PluginInterface.php
@@ -33,6 +33,14 @@ interface PluginInterface
     public function isLicenseJustActivated(): void;
 
     /**
+     * Execute actions when the depended-upon plugin or theme's
+     * status has changed (activated/deactivated).
+     *
+     * @param string[] $pluginFilesOrThemeSlugsWithStatusChange The plugin files or theme slugs with status change
+     */
+    public function dependedUponPluginOrThemeStatusJustChanged(array $pluginFilesOrThemeSlugsWithStatusChange): void;
+
+    /**
      * Plugin name
      */
     public function getPluginName(): string;